v70+ Support, Fixes, Improvements (#12)
* Monster Spawning Stuff

Enable Monsters to spawn on a Map, Regardless of if they are present in the level array or not.

* More Fixes

More Fixes, Should work for All Clients now.

* Better Fix

Credit to @Maya in Discord, works, and is better than mine

* Inverse Teleport (7ark)

Inverse Teleport Player, Credit 7ark

* Inverse Teleporting [7ark]

Inverse Teleporting Client and Crew, credit to 7ark.

* Old Source Code Push

Pushes Source for old Plugin, base for updates

* Fix

What in the

* DANGIT

* Bugfixes

More bugfixes for 1.1.12 update

* missed bool

Missed a bool, indoor enemies may be broken

* Fix Clay Surgeon, Inverse Teleports

* Update CrowdControl.dll

* Monster Spawns Fix

Switch method of adding Enemies back to adding via the Spawn Func, that way we bypass mods that edit the spawn list after us. (NOTE: THIS IS UNTESTED EXCEPT FOR SOLO)

* State checks for timed effects

Fix Timed Effects running when player is dead.

* Recomment Lightning out

Was testing stuff

* Update Version for Game

v70+ Beta Updates

* Live Files, Not Beta

Update is Live, not Beta, Cod3r

* Push v1.1.14

Update Mod to v1.1.14 for v70+

* Comment out unneeded code

This isn't needed in release code.

* Spawn Patching

Works Vanilla, Not with LethalLevelLoader. Need to find a better method to work with mods.

* LLL Compatibility

LethalLevelLoader patches the spawns at 200 priority, 300 seems to work, Tested on March where Radmechs are disabled, and they work again along with others.

* UPDATES:

- Update Namespace names to more fitting names
- Actually Patch spawns on every level run.
- Add Sapsucker and Tulip Snake to Crew

* Update release

Ready for release. v1.1.14

* Quick comment

Comment a print, for future me

* Kidnapper Fox Spawns

Kidnapper Fox. This will need future tweaking, but works fine as is. We spawn 2x Vain Shrouds each time we spawn it, so it should set the nest to the spawn position.

* Remove Kidnapper Fox

Turns out, in v64+ the fox is no longer a network asset, so only visible to the host. May as well remove it then

* More Cleanup

Clean csproj
Remove files that snuck in somehow
Remove unneeded code in CrowdDelegates

* missed this